Answer the second review: finish what the first one started
- Zoom now lives with the other thing Lite remembers about how it looks,
  rather than in the module the component generator owns and every piece
  of the interface imports.
- A step that lands on the size already showing writes nothing, so a held
  key at either end stops writing the same number to disk per repeat.
- Each tile reserves the width of its own widest label: only one action is
  ever offered, so an install row no longer settles for an update row's
  room, nor an update row for an install row's.
- The three id swaps left behind in restart, and the two copies of
  forgetting a shell's agent, go the way the removals already went.
- Restores executable_exists: it existed so its one call site could read
  as one line, and inlining it cost exactly that.
